added th to tversky

+++ lit_models/Vesuvius_Lit_Model.py
@@ -297,7 +297,7 @@ class TverskyLoss(nn.Module):
         # comment out if your model contains a sigmoid or equivalent activation layer
 
         inputs = F.sigmoid(inputs)
-        inputs = (inputs >= 0.4)
+        inputs = (inputs >= 0.4).float()
 
         # flatten label and prediction tensors
         inputs = inputs.view(-1)
